bybit-bot/requirements.txt
httpx>=0.27
pyyaml>=6.0
rich>=13.0
python-dotenv>=1.0
